+++ src/solverpy/benchmark/db/__init__.py
@@ -5,7 +5,7 @@ from .providers.status import Status
 from .providers.loader import Loader, SlashLoader
 
 def default(delfix=None):
-   return DB([Jsons, Solved.Maker(delfix), Status.Maker(delfix)])
+   return DB([Jsons.Maker(), Solved.Maker(delfix=delfix), Status.Maker(delfix=delfix)])
 
 __all__ = ["DB", "Jsons", "JsonsStore", "Solved", "Status", "Loader", 
            "SlashLoader", "default"]