+++ src/solverpy/benchmark/db/provider.py
@@ -1,5 +1,6 @@
+from ...solver.object import SolverPyObj
 
-class Provider:
+class Provider(SolverPyObj):
    """A data provider that stores and/or queries results of tasks."""
 
    def __init__(self, bid, sid, limit=None, store_cached=False):
@@ -9,6 +10,15 @@ class Provider:
       self.store_cached = store_cached
       self._uptodate = False
       "call store for cached results."
+
+   @classmethod
+   def Maker(cls, **kwargs):
+      class MakerMaker(SolverPyObj):
+         def __init__(self):
+            SolverPyObj.__init__(self, myname=cls.__name__, **kwargs)
+         def __call__(self, bid, sid, limit): 
+            return cls(bid, sid, limit, **kwargs)
+      return MakerMaker()
 
    def query(self, task):
       """Return the cached result for `task` or None if not available."""