+++ src/solverpy/solver/plugins/db/outputs.py
@@ -1,26 +1,36 @@
+from typing import TYPE_CHECKING
 import os
 import gzip
 
 from ..decorator import Decorator
 from ....benchmark.path import bids, sids
 
+if TYPE_CHECKING:
+   from ...timedsolver import TimedSolver
+
 NAME = "outputs"
 
 class Outputs(Decorator):
    
-   def __init__(self, flatten=True, compress=True):
+   def __init__(
+      self, 
+      flatten: bool = True, 
+      compress: bool = True
+   ):
+      Decorator.__init__(self, flatten=flatten, compress=compress)
       self._path = bids.dbpath(NAME)
       self._flatten = flatten
       self._compress = compress
    
-   def __repr__(self):
-      return f"{type(self).__name__}({repr(self._flatten)})"
-   
-   def register(self, solver):
+   def register(self, solver : "TimedSolver") -> None:
       solver.decorators.append(self)
       self.solver = solver
    
-   def path(self, instance, strategy):
+   def path(
+      self, 
+      instance: tuple[str,str],
+      strategy: str
+   ) -> str:
       (bid, problem) = instance
       bs = bids.name(bid, limit=self.solver.limits.limit)
       if self._flatten:
@@ -29,11 +39,22 @@ class Outputs(Decorator):
       p = os.path.join(self._path, bs, sids.name(strategy), problem)
       return p
 
-   def finished(self, instance, strategy, output, result):
+   def finished(
+      self, 
+      instance: tuple[str,str], 
+      strategy: str, 
+      output: str, 
+      result: dict
+   ) -> None:
       if output and self.solver.valid(result):
          self.write(instance, strategy, output)
    
-   def write(self, instance, strategy, content):
+   def write(
+      self, 
+      instance: tuple[str,str], 
+      strategy: str, 
+      content: str
+   ) -> None:
       f = self.path(instance, strategy)
       os.makedirs(os.path.dirname(f), exist_ok=True)
       if self._compress: