+++ src/solverpy/solver/plugins/db/sid.py
@@ -1,14 +1,20 @@
+from typing import Any
+
 from ..translator import Translator
 from ....benchmark.path import sids
 
 class Sid(Translator):
    "Strategy ids translator."
 
-   def translate(self, instance, strategy):
+   def translate(
+      self, 
+      instance: Any, 
+      strategy: str
+   ) -> tuple[Any,str]:
       sid = strategy
-      strategy = sids.load(sid)
-      if "@@@" in strategy:
+      strategy0 = sids.load(sid)
+      if "@@@" in strategy0:
          (sid, args) = sids.split(sid)
-         strategy = sids.instatiate(strategy, args)
-      return (instance, strategy)
+         strategy0 = sids.instatiate(strategy0, args)
+      return (instance, strategy0)