+++ src/solverpy/solver/plugins/shell/memory.py
@@ -5,6 +5,7 @@ ULIMIT_CMD = "ulimit -Sv %d"
 class Memory(Decorator):
    
    def __init__(self, giga=4):
+      Decorator.__init__(self, giga=giga)
       self.prefix = ULIMIT_CMD % int(giga * 1000000)
    
    def decorate(self, cmd, instance, strategy):