+++ src/solverpy/solver/plugins/shell/timeout.py
@@ -5,12 +5,10 @@ TIMEOUT_CMD = "timeout --kill-after=15 --foreground %s"
 class Timeout(Decorator):
    
    def __init__(self, timeout):
+      Decorator.__init__(self, timeout=timeout)
       self.timeout = timeout
       self.prefix = TIMEOUT_CMD % timeout 
    
-   def __repr__(self):
-      return f"{type(self).__name__}({repr(self.timeout)})"
-
    def register(self, solver):
       self.solver = solver
       solver.decorators.insert(0, self)