+++ src/solverpy/solver/pluginsolver.py
@@ -1,20 +1,28 @@
+from typing import Any
 import logging
+
 from .solver import Solver
+from .plugins.plugin import Plugin
+from .plugins.decorator import Decorator
+from .plugins.translator import Translator
 
 logger = logging.getLogger(__name__)
   
 class PluginSolver(Solver):
 
-   def __init__(self, plugins=[]):
-      self.decorators = []
-      self.translators = []
+   def __init__(self, plugins : list[Plugin] = []):
+      self.decorators : list[Decorator] = []
+      self.translators : list[Translator] = []
       self.init(plugins)
 
-   def __repr__(self):
-      plgs = ",".join(repr(p) for p in self.decorators+self.translators)
-      return plgs
+   def represent(self):
+      return dict(
+         cls=self.name,
+         decorators=[repr(x) for x in self.decorators],
+         translators=[repr(x) for x in self.translators],
+      ) 
 
-   def solve(self, instance, strategy):
+   def solve(self, instance : Any, strategy : Any) -> dict:
       (output, result) = super().solve(instance, strategy)
       self.update(instance, strategy, output, result)
       if not self.valid(result):
@@ -27,25 +35,25 @@ class PluginSolver(Solver):
       return result
    
    # plugins initization
-   def init(self, plugins):
+   def init(self, plugins : list[Plugin]) -> None:
       for plugin in plugins:
          plugin.register(self)
    
    # decorators
    # decorate(self, command, input)
-   def decorate(self, cmd, instance, strategy):
+   def decorate(self, cmd : str, instance : Any, strategy : Any):
       for plugin in self.decorators:
          cmd = plugin.decorate(cmd, instance, strategy)
       return cmd
    
-   def update(self, instance, strategy, output, result):
+   def update(self, instance : Any, strategy : Any, output : str, result : dict):
       for plugin in self.decorators:
          plugin.update(instance, strategy, output, result)
       for plugin in self.decorators:
          plugin.finished(instance, strategy, output, result)
 
    # translators
-   def translate(self, instance, strategy):
+   def translate(self, instance : Any, strategy : Any):
       for plugin in self.translators:
          (instance, strategy) = plugin.translate(instance, strategy)
       return (instance, strategy)